extract out test utils to remove enzyme dependency
