+++ packages/airglow-plugin-console/test/int/console.plugin.test.js
@@ -13,7 +13,7 @@ governing permissions and limitations under the License.
 /* eslint-disable no-console */
 
 import React from 'react';
-import { renderAirglow } from 'airglow';
+import { renderAirglow } from '@airglow/test-utils';
 import ReactStore from '@airglow/store-react';
 import { call } from '@airglow/reducers';
 import plugin from '../../src/index';
