+++ packages/airglow-plugin-devtools/package.json
@@ -23,6 +23,7 @@
   },
   "devDependencies": {
     "@airglow/store-redux": "^0.5.2",
+    "@airglow/test-utils": "^0.0.1",
     "@babel/cli": "^7.4.4"
   },
   "resolutions": {
