+++ packages/airglow-plugin-prefab/test/int/plugin.test.js
@@ -11,7 +11,7 @@ governing permissions and limitations under the License.
 */
 
 import React from 'react';
-import { renderAirglow } from 'airglow';
+import { renderAirglow } from '@airglow/test-utils';
 import ReactStore, { connect } from '@airglow/store-react';
 import prefab from '@airglow/prefab';
