+++ packages/airglow-plugin-saga/test/int/plugin.test.js
@@ -12,7 +12,7 @@ governing permissions and limitations under the License.
 
 import React from 'react';
 import ReduxStore, { connect } from '@airglow/store-redux';
-import { renderAirglow } from 'airglow';
+import { renderAirglow } from '@airglow/test-utils';
 import { takeEvery } from 'redux-saga/effects';
 import plugin, { SAGA } from '../../src/index';
