+++ packages/airglow-plugin-test/test/int/plugin.test.js
@@ -11,7 +11,7 @@ governing permissions and limitations under the License.
 */
 
 import React from 'react';
-import { renderAirglow } from 'airglow';
+import { renderAirglow } from '@airglow/test-utils';
 import ReactStore from '@airglow/store-react';
 import plugin, { lastDispatch, dispatchList } from '../../src/index';
