+++ packages/airglow-store-redux/package.json
@@ -22,6 +22,7 @@
   },
   "dependencies": {
     "@airglow/reducers": "^0.5.2",
+    "@airglow/test-utils": "^0.0.1",
     "airglow": "^0.5.2"
   },
   "devDependencies": {
