+++ packages/airglow-test-utils/src/index.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 Adobe. All rights reserved.
+Copyright 2020 Adobe. All rights reserved.
 This file is licensed to you under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License. You may obtain a copy
 of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -10,9 +10,9 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
+import Airglow from 'airglow';
 import React from 'react';
 import { mount } from 'enzyme';
-import Airglow from '../airglow';
 
 export const renderAirglow = (children, props) => {
   let airglow;
