+++ packages/airglow/package.json
@@ -23,6 +23,7 @@
     "ramda": "^0.26.1"
   },
   "devDependencies": {
+    "@airglow/test-utils": "^0.0.1",
     "@babel/cli": "^7.4.4"
   },
   "scripts": {
