+++ packages/airglow/src/index.js
@@ -17,5 +17,3 @@ export default Airglow;
 export * from './types';
 
 export { AirglowProvider, AirglowWrapper };
-
-export { fakeStore, renderAirglow } from './util/test.util';
