+++ packages/airglow/test/src/airglow.spec.js
@@ -14,10 +14,11 @@ import React from 'react';
 import sinon from 'sinon';
 import { mount, shallow } from 'enzyme';
 
+import { fakeStore } from '@airglow/test-utils';
+
 import Airglow from '../../src/airglow';
 import * as engine from '../../src/engine';
 import { COMPONENT, HOC } from '../../src/types';
-import { fakeStore } from '../../src/util/test.util';
 
 let bootstrapSpy;
 let pluginSpy;
